refactor: API 경로 수정 및 도메인에 맞게 로직 분리 (#177)
