+++ src/api/taste-test/index.ts
@@ -1,13 +1,8 @@
 import { API_PATHS } from '@/constants/apiPaths'
-import type { AnswerType, TasteTestProfile } from '@/types/tasteTypes'
+import type { AnswerType } from '@/types/tasteTypes'
 import { axiosInstance } from '@/utils/axios'
 
 export const tasteTestApi = {
-  getTasteTestProfile: async (): Promise<TasteTestProfile> => {
-    const { data } = await axiosInstance.get(API_PATHS.TASTE_TEST.PROFILE)
-    return data
-  },
-
   //테스트 문항 가져오기 함수
   getTestQuestion: async () => {
     const { data } = await axiosInstance.get(API_PATHS.TASTE_TEST.QUESTIONS)
