--- src/hooks/taste-history/useTasteHistory.ts
@@ -1,11 +1,0 @@
-import { feedbackApi } from '@/api/feedback/index'
-import { useQuery } from '@tanstack/react-query'
-import { commonQueryOptions } from '@/hooks/home/useProduct'
-
-export const useTasteHistory = () => {
-  return useQuery({
-    queryKey: ['tasteHistory'],
-    queryFn: feedbackApi.tastingHistory,
-    ...commonQueryOptions,
-  })
-}
